Struct tokio_threadpool::pool::backup::Backup

pub(crate) struct Backup {
    handoff: UnsafeCell<Option<WorkerId>>,
    state: AtomicUsize,
    next_sleeper: UnsafeCell<BackupId>,
    park: DefaultPark,
}

State associated with a thread in the thread pool.

The pool manages a number of threads. Some of those threads are considered “primary” threads and process the work queue. When a task being run on a primary thread enters a blocking context, the responsibility of processing the work queue must be handed off to another thread. This is done by first checking for idle threads on the backup stack. If one is found, the worker token (WorkerId) is handed off to that running thread. If none are found, a new thread is spawned.

This state manages the exchange. A thread that is idle, not assigned to a work queue, sits around for a specified amount of time. When the worker token is handed off, it is first stored in handoff. The backup thread is then signaled. At this point, the backup thread wakes up from sleep and reads handoff. At that point, it has been promoted to a primary thread and will begin processing inbound work on the work queue.

The name Backup isn’t really great for what the type does, but I have not come up with a better name… Maybe it should just be named Thread.

Fields
handoff: UnsafeCell<Option<WorkerId>>

Worker ID that is being handed to this thread.

state: AtomicUsize

Thread state.

This tracks:

Is queued flag

If the pool is shutting down.

If the thread is running

next_sleeper: UnsafeCell<BackupId>

Next entry in the Treiber stack.

park: DefaultPark

Used to put the thread to sleep

Implementations

impl Backup

pub fn new() -> Backup

pub fn start(&self, worker_id: &WorkerId)

Called when the thread is starting

pub fn is_running(&self) -> bool

pub fn worker_handoff(&self, worker_id: WorkerId) -> bool

Hands off the worker to a thread.

Returns true if the thread needs to be spawned.

pub fn signal_stop(&self)

Terminate the worker

pub fn release(&self)

Release the worker

pub fn wait_for_handoff(&self, timeout: Option<Duration>) -> Handoff

Wait for a worker handoff

pub fn is_pushed(&self) -> bool

pub fn set_pushed(&self, ordering: Ordering)

pub fn next_sleeper(&self) -> BackupId

pub fn set_next_sleeper(&self, val: BackupId)