Struct tokio_threadpool::pool::Pool

pub(crate) struct Pool {
    pub state: CachePadded<AtomicUsize>,
    sleep_stack: CachePadded<Stack>,
    pub workers: Arc<[Entry]>,
    pub queue: Arc<Injector<Arc<Task>>>,
    pub trigger: Weak<ShutdownTrigger>,
    backup: Box<[Backup]>,
    pub backup_stack: BackupStack,
    blocking: Blocking,
    pub config: Config,
}

Fields
state: CachePadded<AtomicUsize>sleep_stack: CachePadded<Stack>workers: Arc<[Entry]>queue: Arc<Injector<Arc<Task>>>trigger: Weak<ShutdownTrigger>backup: Box<[Backup]>backup_stack: BackupStackblocking: Blockingconfig: Config
Implementations

impl Pool

pub fn new(
workers: Arc<[Entry]>,
trigger: Weak<ShutdownTrigger>,
max_blocking: usize,
config: Config,
queue: Arc<Injector<Arc<Task>>>
) -> Pool

Create a new Pool

pub fn shutdown(&self, now: bool, purge_queue: bool)

Start shutting down the pool. This means that no new futures will be accepted.

pub fn push_sleeper(&self, idx: usize) -> Result<(), ()>

Called by Worker as it tries to enter a sleeping state. Before it sleeps, it must push itself onto the sleep stack. This enables other threads to see it when signaling work.

pub fn terminate_sleeping_workers(&self)

pub fn poll_blocking_capacity(
&self,
task: &Arc<Task>
) -> Poll<(), BlockingError>

pub fn submit(&self, task: Arc<Task>, pool: &Arc<Pool>)

Submit a task to the scheduler.

Called from either inside or outside of the scheduler. If currently on the scheduler, then a fast path is taken.

pub fn submit_external(&self, task: Arc<Task>, pool: &Arc<Pool>)

Submit a task to the scheduler from off worker

Called from outside of the scheduler, this function is how new tasks enter the system.

pub fn release_backup(&self, backup_id: BackupId) -> Result<(), ()>

pub fn notify_blocking_task(&self, pool: &Arc<Pool>)

pub fn spawn_thread(&self, id: WorkerId, pool: &Arc<Pool>)

Provision a thread to run a worker

pub fn signal_work(&self, pool: &Arc<Pool>)

If there are any other workers currently relaxing, signal them that work is available so that they can try to find more work to process.

pub fn rand_usize(&self) -> usize

Generates a random number

Uses a thread-local random number generator based on XorShift.