Struct tokio_threadpool::task::blocking_state::BlockingState

pub(crate) struct BlockingState(usize);

State tracking task level state to support blocking.

This tracks two separate flags.

a) If the task is queued in the pending blocking channel. This prevents double queuing (which would break the linked list).

b) If the task has been allocated capacity to block.

Implementations

impl BlockingState

pub fn new() -> BlockingState

Create a new, default, BlockingState.

pub fn is_queued(&self) -> bool

Returns true if the state represents the associated task being queued in the pending blocking capacity channel

pub fn toggle_queued(state: &AtomicUsize, ordering: Ordering) -> BlockingState

Toggle the queued flag

Returns the state before the flag has been toggled.

pub fn is_allocated(&self) -> bool

Returns true if the state represents the associated task having been allocated capacity to block.

pub fn consume_allocation(state: &AtomicUsize, ordering: Ordering) -> CanBlock

Atomically consume the capacity allocation and return if the allocation was present.

If this returns true, then the task has the ability to block for the duration of the poll.

pub fn notify_blocking(state: &AtomicUsize, ordering: Ordering)