Struct tokio_threadpool::task::Task

pub(crate) struct Task {
    state: AtomicUsize,
    blocking: AtomicUsize,
    next_blocking: AtomicPtr<Task>,
    pub reg_worker: Cell<Option<u32>>,
    pub reg_index: Cell<usize>,
    future: UnsafeCell<Option<Spawn<Box<dyn Future<Item = (), Error = ()> + Send + 'static>>>>,
}

Harness around a future.

This also behaves as a node in the inbound work queue and the blocking queue.

Fields
state: AtomicUsize

Task lifecycle state

blocking: AtomicUsize

Task blocking related state

next_blocking: AtomicPtr<Task>

Next pointer in the queue of tasks pending blocking capacity.

reg_worker: Cell<Option<u32>>

ID of the worker that polled this task first.

This field can be a Cell because it’s only accessed by the worker thread that is executing the task.

The worker ID is represented by a u32 rather than usize in order to save some space on 64-bit platforms.

reg_index: Cell<usize>

The key associated with this task in the Slab it was registered in.

This field can be a Cell because it’s only accessed by the worker thread that has registered the task.

future: UnsafeCell<Option<Spawn<Box<dyn Future<Item = (), Error = ()> + Send + 'static>>>>

Store the future at the head of the struct

The future is dropped immediately when it transitions to Complete

Implementations

impl Task

pub fn new(
future: Box<dyn Future<Item = (), Error = ()> + Send + 'static>
) -> Task

Create a new Task as a harness for future.

fn stub() -> Task

Create a fake Task to be used as part of the intrusive mpsc channel algorithm.

pub fn run(&self, unpark: &Arc<Notifier>) -> Run

Execute the task returning Run::Schedule if the task needs to be scheduled again.

pub fn abort(&self)

Aborts this task.

This is called when the threadpool shuts down and the task has already beed polled but not completed.

pub fn notify(me: Arc<Task>, pool: &Arc<Pool>)

Notify the task

pub fn notify_blocking(me: Arc<Task>, pool: &Arc<Pool>)

Notify the task it has been allocated blocking capacity

pub fn schedule(&self) -> bool

Transition the task state to scheduled.

Returns true if the caller is permitted to schedule the task.

pub fn consume_blocking_allocation(&self) -> CanBlock

Consumes any allocated capacity to block.

Returns true if capacity was allocated, false otherwise.

fn drop_future(&self)

Drop the future

This must only be called by the thread that successfully transitioned the future state to Running.