Crate tokio_threadpool

A work-stealing based thread pool for executing futures.

Modules

park
Thread parking utilities.

Structs

BlockingError
Error raised by blocking.

Builder
Builds a thread pool with custom configuration values.

Sender
Submit futures to the associated thread pool for execution.

Shutdown
Future that resolves when the thread pool is shutdown.

ThreadPool
Work-stealing based thread pool for executing futures.

Worker
Thread worker

Functions

blocking
Enter a blocking section of code.