Struct tokio_threadpool::worker::entry::WorkerEntry

pub(crate) struct WorkerEntry {
    pub state: CachePadded<AtomicUsize>,
    next_sleeper: UnsafeCell<usize>,
    pub worker: Worker<Arc<Task>>,
    stealer: Stealer<Arc<Task>>,
    park: UnsafeCell<Option<Box<dyn Park<Unpark = Box<dyn Unpark>, Error = ()> + Send>>>,
    unpark: UnsafeCell<Option<Box<dyn Unpark>>>,
    running_tasks: UnsafeCell<Slab<Arc<Task>>>,
    remotely_completed_tasks: SegQueue<Arc<Task>>,
    needs_drain: AtomicBool,
}

Fields
state: CachePadded<AtomicUsize>next_sleeper: UnsafeCell<usize>worker: Worker<Arc<Task>>stealer: Stealer<Arc<Task>>park: UnsafeCell<Option<Box<dyn Park<Unpark = Box<dyn Unpark>, Error = ()> + Send>>>unpark: UnsafeCell<Option<Box<dyn Unpark>>>running_tasks: UnsafeCell<Slab<Arc<Task>>>remotely_completed_tasks: SegQueue<Arc<Task>>needs_drain: AtomicBool
Implementations

impl WorkerEntry

pub fn new(
park: Box<dyn Park<Unpark = Box<dyn Unpark>, Error = ()> + Send>,
unpark: Box<dyn Unpark>
) -> Self

pub fn fetch_unset_pushed(&self, ordering: Ordering) -> State

Atomically unset the pushed flag.

Return

The state before the push flag is unset.

Ordering

The specified ordering is established on the entry’s state variable.

pub fn submit_internal(&self, task: Arc<Task>)

Submit a task to this worker while currently on the same thread that is running the worker.

pub fn notify(&self, state: State) -> bool

Notifies the worker and returns false if it needs to be spawned.

Ordering

The state must have been obtained with an Acquire ordering.

pub fn signal_stop(&self, state: State)

Signals to the worker that it should stop

state is the last observed state for the worker. This allows skipping the initial load from the state atomic.

Return

Returns Ok when the worker was successfully signaled.

Returns Err if the worker has already terminated.

pub fn pop_task(&self) -> Option<Arc<Task>>

Pop a task

This must only be called by the thread that owns the worker entry. This function is not Sync.

pub fn steal_tasks(&self, dest: &Self) -> Steal<Arc<Task>>

Steal tasks

This is called by other workers to steal a task for processing. This function is Sync.

At the same time, this method steals some additional tasks and moves them into dest in order to balance the work distribution among workers.

pub fn drain_tasks(&self)

Drain (and drop) all tasks that are queued for work.

This is called when the pool is shutting down.

pub fn park(&self)

Parks the worker thread.

pub fn park_timeout(&self, duration: Duration)

Parks the worker thread for at most duration.

pub fn unpark(&self)

Unparks the worker thread.

pub fn register_task(&self, task: &Arc<Task>)

Registers a task in this worker.

Called when the task is being polled for the first time.

pub fn unregister_task(&self, task: Arc<Task>)

Unregisters a task from this worker.

Called when the task is completed and was previously registered in this worker.

pub fn remotely_complete_task(&self, task: Arc<Task>)

Unregisters a task from this worker.

Called when the task is completed by another worker and was previously registered in this worker.

pub fn shutdown(&self)

Drops the remaining incomplete tasks and the parker associated with this worker.

This function is called by the shutdown trigger.

fn drain_remotely_completed_tasks(&self)

Drains the remotely_completed_tasks queue and removes tasks from running_tasks.

pub fn push_internal(&self, task: Arc<Task>)

pub fn next_sleeper(&self) -> usize

pub fn set_next_sleeper(&self, val: usize)